add mongoose & add top100 playlist
* add mongoose db (will be removed)
* add top100 playlist (based on crawl from melon.com)
