+++ pages/new/service/index/music/index.vue
@@ -67,8 +67,8 @@ export default {
       const songId = location.search.split("?")[1].split("=")[1].split("&")[0];
       const this_out = this;
       this_out.searchKey = location.search.split("&")[1].split("=")[1];
-      axios.post("/api/songDetail", {
-        reqService: "BASE_WEB_ARCHITEC",
+      axios.post("/api/song/detail", {
+        reqService: "BASE_WEB_ARCHITEC_PLY_MINI",
         songId
       })
       .then(({data}) => {
@@ -98,13 +98,6 @@ export default {
               $(".selectedObjF").css("height", $(widnow).height());
           }
           const songId = location.search.split("?")[1].split("=")[1];
-          // axios.post("/api/songDetail", {
-          //   reqService: "BASE_WEB_ARCHITEC",
-          //   songId
-          // })
-          // .then(({data}) => {
-          //   console.log(data);
-          // })
 
 
           const colorThief = new ColorThief();
